:bento: Added plane controller and dummy material
The dummy material is intentionally included with transparent renderer
mode so that final build expects transparent renders. This allows me to
switch between modes during runtime; which is used for TrackOnFloor
script.

We can use the new plane controller static script so we can
activate/deactivate plane manager and current detected planes in the
real world.